fix [NSTableView setDropRow:] types
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/branches/gnustep_testplant_branch@40195 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSTableView.m
@@ -5623,7 +5623,7 @@ This method is deprecated, use -columnIndexesInRect:. */
   return [image autorelease];
 }
 
-- (void) setDropRow: (int)row
+- (void) setDropRow: (NSInteger)row
       dropOperation: (NSTableViewDropOperation)operation
 {
   if (row < -1 || row > _numberOfRows
